Restructured the code so that I can run from my windows box while developing and we can still run the same code base from the site5 host when it's deployed.  Verified it works on both.
Structure:
app/__init__.py - This is the python package file.  app will contain the actuall web app code.

app/views.py - Responsible for routing hillview website urls to the appropriate code.  For now it renders . and /index to the same message I've had on there for a while.

run.py - You can run this from windows and it will launch a local server to see your development changes.

index.fcgi - Updated with the minimal code needed for the Site5 host to run the code in app/